requirements.txt
flax==0.7.1
jax==0.4.11
jaxlib==0.4.11
matplotlib
ml-dtypes==0.2.0
mypy
omegaconf==2.3.0
optax==0.1.5
orbax-checkpoint==0.3.5
pillow
pre-commit>=3.3.3
torch>=2.0.1
torchvision>=0.15.2
